Add Jenkinsfile
At the moment the jenkinsfile hasn't split stuff out into helpers or
does anything _well_ due to using the old service pipeline stuff. It
still works though so we should use it given that fact then we can get
TCS deployed out quickly and trial it.

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>

Update Jenkinsfile to allow for branches

We need to check and build the branch and announce to GitHub so lets add
some more fluff to this file to do that.

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>

Update try/catch/finally block

We need to allow the error to be passed

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>

Add pending status for PR kick off.

This is the last part of the puzzle for the GitHub status editing

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>

Final updates to Jenkinsfile

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>

Update Jenkinsfile

Remove calls as they're done automatically now

Signed-off-by: Chris M <chris.mills@hee.nhs.uk>