Add unit tests for placement added/amended dates
Update PlacementServiceImpl to provide a Clock object to
LocalDateTime.now() to allow the clock to be mocked or replaced with a
fixed instant clock for testing purposes.

Update PlacementServiceImplTest to add unit tests to check that added
and amended dates are added when expected.

Update DateTimeFormatConfiguration to add a Clock bean.