Update Post Grade, Site and Speciality DTOs
Update PostGradeDTO, PostSiteDTO and PostSpecialityDTO to add a
constructor which accepts the common fields, this removes the need to
call the same setters whenever the DTO is initialized.

Update tcs-api's pom.xml to bump the minor version.

Update tcs-client's pom.xml to update the tcs-api dependency and bump
its own minor version.