Fix potential NullPointerException
When there are no existing fundings on a post a NullPointerException
will be thrown when trying to add new PostFundingDTOs to the PostDTO.

Update PostDTO to add an addFunding() method which handles this
scenario.

Update PostServiceImpl to use addFunding() instead of getting the
existing collection, which may be null, and adding to it.