Refactor DTOs, use Lombok getter, setter, toString
Lombok can be used to generate a lot of the boilerplate code, such as
constructors and getter, setter, equals and toString methods.
Update the DTOs to use as much boiilerplate generation as possible.

equals() and hashCode() should not be generated at this point as most of
the DTOs use a sub-selection of fields rather than all fields.