fix: Resolve reflection sonar vulnerabilities
Update CommentServiceImpl to use the new Comment entity after the
saveAndFlush operation, rather than returning the original.

Update ProgrammeMembershipResource to return the ID of the entity in the
header instead of the whole input entity object.

Update RotationPostResource to return a message when the request body is
empty instead of returning the request body itself.

TISNEW-3999