test: Disable failing integration tests
The DocumentResource integration tests fail unless `eu-west-2` is
provided as the AWS region. This is due to the test using a real S3
bucket in the London region.

As there should be no reliance on real resources in this way, these
tests should be disabled and replaced with mocks until a full S3 mock
can be implemented.

TIS21-2116