chore: split deployments so failures don't cascade
When deploying all module with the same command a failure in earlier
modules will cause a later failure. This can not be mitigated by
``--fail-at-end` due to dependencies between the modules.

Without this split then it would never be possible to publish a new
version of `tcs-client` or `tcs-persistence` unless `tcs-api` is also
bumped and published.

TIS21-2495