feat(ntn): populate NTN in PM data
The programme membership DTO has a legacy TrainingNumber field which
does not get popualted.
Update the TrainingNumberService to allow calculation and population of
TrainingNumbers in to a list of programme memberships.
Update the ProgrammeMembershipService to call the new TrainingNumber
population when fetching PMs to be displayed in Admins UI.

TIS21-6375
TIS21-6378