fix:remove the outer programmeMembership status filter
Previously this filter didn't work, and last iteration we added it back,
which resulted in some users who don't have programme admin/observers roles
not able to see person records which don't have any active programme memberships.
So we decided to take it off as a patch to resume this functionality to be
the same as before.

However, the programme membership filter doesn't work well and has to be reviewed
in the future.

TIS21-6359