fix: getting current posts for a programme
The status field can no longer be maintained.
This fixes one of many endpoints that use the status field.
As an interim fix, the `fundingStatus` is used as it is maintained.

We have not identified a preferred solution.
Therefore, I am not commiting an additional, temporary test.
To replicate the bug tests would use a deprecated field.

TIS21-7140: Post searches using legacy status